+++ src/Controllers/VenueController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Controllers;
 
+use App\Helpers\Auth;
 use App\Models\VenueModel;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
@@ -29,6 +30,9 @@ class VenueController
 
     public function create(Request $request, Response $response): Response
     {
+        if ($redirect = Auth::requireAdmin($response, $this->basePath)) {
+            return $redirect;
+        }
         $html = $this->twig->render('venue/create.html.twig', [
             'base_path' => $this->basePath,
         ]);
@@ -38,6 +42,9 @@ class VenueController
 
     public function store(Request $request, Response $response): Response
     {
+        if ($redirect = Auth::requireAdmin($response, $this->basePath)) {
+            return $redirect;
+        }
         $data = $request->getParsedBody();
         $this->venueModel->create(
             (string) ($data['name'] ?? ''),
@@ -51,6 +58,9 @@ class VenueController
 
     public function edit(Request $request, Response $response, array $args): Response
     {
+        if ($redirect = Auth::requireAdmin($response, $this->basePath)) {
+            return $redirect;
+        }
         $venue = $this->venueModel->getById((int) $args['id']);
 
         if (!$venue) {
@@ -67,6 +77,9 @@ class VenueController
 
     public function update(Request $request, Response $response, array $args): Response
     {
+        if ($redirect = Auth::requireAdmin($response, $this->basePath)) {
+            return $redirect;
+        }
         $id    = (int) $args['id'];
         $data  = $request->getParsedBody();
         $venue = $this->venueModel->load($id);
@@ -85,6 +98,9 @@ class VenueController
 
     public function destroy(Request $request, Response $response, array $args): Response
     {
+        if ($redirect = Auth::requireAdmin($response, $this->basePath)) {
+            return $redirect;
+        }
         $venue = $this->venueModel->load((int) $args['id']);
         if ($venue->id) {
             $this->venueModel->delete($venue);
